x86_64: Fix encrypted and signature packet layout
Signed and encrypted messages contain the size of the signature,
plaintext, and ciphertext, respectively. Right the corresponding use the
data type size_t, since they all contain sizes. However, this integer
type has no guarantees regarding its size. Typically, it is 32 bits on
32-bit machines and 64 bits on 64-bit architectures. This can cause
verification and decryption failures, when machines of different
architectures communicate with each other. This commit fixes the issue
by replacing usage of size_t with integers of specific widths.

+++ src/Principal.cc
@@ -121,7 +121,7 @@ bool Principal::verify_signature(const char *src, unsigned src_len,
   INCR_OP(num_sig_ver);
   START_CC(sig_ver_cycles);
 
-  size_t signature_len;
+  uint32_t signature_len;
   memcpy(&signature_len, sig, sizeof(signature_len));
   sig += sizeof(signature_len);
   if (signature_len + sizeof(signature_len) > sig_size()) {
@@ -138,9 +138,9 @@ bool Principal::verify_signature(const char *src, unsigned src_len,
   return ret;
 }
 
-unsigned Principal::encrypt(const char *src, unsigned src_len, char *dst,
+unsigned Principal::encrypt(const char *src, uint32_t src_len, char *dst,
                             unsigned dst_len) {
-  unsigned ciphertext_len = pkey->size();
+  uint32_t ciphertext_len = pkey->size();
   unsigned total_len =
       ciphertext_len + sizeof(src_len) + sizeof(ciphertext_len);
   if (dst_len < total_len) {
